+++ src/main/java/com/elcafe/RestaurantDeliveryApplication.java
@@ -6,12 +6,14 @@ import org.springframework.cache.annotation.EnableCaching;
 import org.springframework.cloud.openfeign.EnableFeignClients;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 import org.springframework.scheduling.annotation.EnableAsync;
+import org.springframework.scheduling.annotation.EnableScheduling;
 
 @SpringBootApplication
 @EnableJpaAuditing
 @EnableCaching
 @EnableFeignClients
 @EnableAsync
+@EnableScheduling
 public class RestaurantDeliveryApplication {
 
     public static void main(String[] args) {